requirements.txt
# GitHub API
PyGithub==2.1.1

# Environment variables
python-dotenv==1.0.0
